criando endpoins e services de roles e users, e ajustes na buscas genericas

+++ projeto/backend/csharp_p2/src/Modules/Domain/Roles/RolesService.cs
@@ -1,5 +1,6 @@
 using csharp_p2.src.Modules.Entities;
 using csharp_p2.src.Shared.DTOs.Roles;
+using Shared.Exceptions;
 
 namespace csharp_p2.src.Modules.Domain.Roles;
 
@@ -13,11 +14,14 @@ public interface IRolesService {
 
 public class RolesService : IRolesService {
   private readonly IGenericEntityRepository<Role> _roleRepository;
+  private readonly IGenericEntityRepository<User> _userRepository;
 
-  public RolesService(IGenericEntityRepository<Role> roleRepository) {
+  public RolesService(IGenericEntityRepository<Role> roleRepository, IGenericEntityRepository<User> userRepository) {
     _roleRepository = roleRepository;
+    _userRepository = userRepository;
   }
 
+  //!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!GETS
   public async Task<IEnumerable<Role>> GetAllRolesAsync() {
     var roles = await _roleRepository.GetAllAsync();
     return roles;
@@ -28,9 +32,15 @@ public class RolesService : IRolesService {
     return _roleRepository.GetByIdAsync(id);
   }
 
+  //!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!CREATE
   public async Task<Role> CreateRoleAsync(RoleDto dto) {
     await using var trx = await _roleRepository.GetContext().Database.BeginTransactionAsync();
     try {
+      var existingRole = await _roleRepository.FindOneWithPredicateAsync((r) =>
+        r.Description.Equals(dto.Description, StringComparison.OrdinalIgnoreCase)
+      );
+      if (existingRole != null) throw new CustomError("Role already exists");
+
       var role = new Role {
         Description = dto.Description
       };
@@ -43,22 +53,29 @@ public class RolesService : IRolesService {
     }
   }
 
+  //!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!UPDATE
   public async Task<Role> UpdateRoleAsync(long id, RoleDto dto) {
     var role = await _roleRepository.GetByIdAsync(id);
-    if (role == null) throw new Exception("Role not found");
+    if (role == null) throw new CustomError("Role not found");
 
     role.Description = dto.Description;
 
     return await _roleRepository.UpdateAsync(role)
               .ContinueWith(t => {
                 if (t.Result) return role;
-                throw new Exception("Erro ao atualizar Role");
+                throw new Exception("Erro ao atualizar Role: " + t.Exception?.Message);
               });
   }
 
+  //!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!DELETE
   public async Task<bool> DeleteRoleAsync(long id) {
     await using var trx = await _roleRepository.GetContext().Database.BeginTransactionAsync();
     try {
+      var usersWithRole = await _userRepository.FindOneWithPredicateAsync(u => u.RoleId == id);
+      if (usersWithRole != null) {
+        throw new CustomError("Cannot delete role with associated users.");
+      }
+
       var deleted = await _roleRepository.DeleteAsync(id);
       await trx.CommitAsync();
       return deleted;
